neighbornook_backend/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.6.2
crispy-bootstrap5==2024.2
dj-database-url==2.2.0
Django==4.2.14
django-bootstrap-v5==1.0.11
django-bootstrap5==24.2
django-cors-headers==4.4.0
django-crispy-forms==2.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
environs==11.0.0
gunicorn==22.0.0
marshmallow==3.21.3
packaging==24.1
pillow==10.3.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
soupsieve==2.5
sqlparse==0.5.0
typing_extensions==4.12.2
whitenoise==6.7.0
